requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
colorama==0.4.4
cryptography==35.0.0
Django==3.2.9
django-cors-headers==3.10.0
django-rest-knox==4.1.0
django-rest-passwordreset==1.2.1
djangorestframework==3.12.4
filelock==3.3.2
future==0.18.2
gevent==21.8.0
greenlet==1.1.2
idna==3.3
Pillow==8.4.0
pycparser==2.20
PySocks==1.7.1
pytz==2021.3
requests==2.26.0
requests-file==1.5.1
six==1.16.0
sqlparse==0.4.2
tldextract==3.1.2
typing-extensions==3.10.0.2
urllib3==1.26.7
zope.event==4.5.0
zope.interface==5.4.0
